tsa-infrastructure/requirements.txt
aws-cdk-lib>=2.200.0
constructs>=10.0.0
aws-cdk.aws-amplify-alpha

# Install shared layer as editable package
-e ./layers/tsa-shared-utilities/python